arbeider/requirements.txt
Django>=2.0.0,<3
channels>=2.0.2
channels_redis~=2.0
django-extensions
drf-access-policy

django-cors-headers
django-filter
djangorestframework
drf-access-policy



scipy
graphene-django

pandas-profiling
zarr
dask[complete]==2.9.2
lz4
numcodecs
xarray
seaborn
blosc
dask-image